requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow==10.0.1
py2exe==0.13.0.1
sv-ttk==2.6.0
ttkthemes==3.2.2
